requirements.txt
PyQt5
google-api-python-client
youtube-transcript-api
python-dotenv
